tests/wpt/web-platform-tests/workers/WorkerGlobalScope_ErrorEvent_lineno.htm

<!DOCTYPE html>
<title> WorkerGlobalScope onerror event handler argument: line </title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id=log></div>
<script>
async_test(function(t) {
  var worker = new Worker('./support/ErrorEvent.js');
  worker.onmessage = t.step_func_done(function(e) {
    assert_equals(e.data.lineno, 3);
  });
  worker.postMessage("Error Message");
});
</script>
